update: moved function to reducer file for jest tests
